`index.html`: Add version as query parameter (#13)
Add the current version of the application as a query parameter
to the `/submit` and `/validate` actions.

Now, if the user hits the "submit" button to submit a review, they
essentially make a `POST` request to either `/submit?version=2.0.1`
or `/submit?version=0.1.1`, depending on the version of the app
that they're running (0.1.1 for no emojis and 2.0.1 for emojis).

This `version` query parameter will help the Istio ingressgateway
to route each `POST` request to the correct corresponding Pod. In
a few words, we avoid sending `POST` requests that were meant for
one version of the app to the wrong Pod.